requirements.txt
Django==2.0.3
django-bootstrap4==0.0.6
django-filter==1.1.0
djangorestframework==3.7.7
Markdown==2.6.11
pytz==2018.3
python-dateutil==2.7.3
django-dotenv==1.4.1
django-getenv==1.3.1
dj-database-url==0.4.2
pyldap==2.4.28
django-auth-ldap==1.2.10
